chore: move execution details components to corresponding pages
