--- src/components/molecules/ExecutionDetails/index.ts
@@ -1,2 +1,0 @@
-export {default as TestExecutionDetailsTabs} from './TestExecutionDetails/TestExecutionDetailsTabs';
-export {default as TestSuiteExecutionDetailsTabs} from './TestSuiteExecutionDetails/TestSuiteExecutionDetailsTabs';
