+++ src/components/pages/TestSuites/TestSuiteDetails/TestSuiteExecution/TestSuiteExecutionDrawer.tsx
@@ -1,18 +1,18 @@
 import {FC} from 'react';
 
-import ExecutionDrawer from 'src/components/organisms/EntityDetails/ExecutionDrawer';
-
-import {TestSuiteExecutionDetailsTabs} from '@molecules';
+import ExecutionDrawer from '@organisms/EntityDetails/ExecutionDrawer';
 
 import ExecutionDrawerHeader from '@organisms/EntityDetails/ExecutionDrawer/ExecutionDrawerHeader';
 
 import {useExecutionDetailsPick} from '@store/executionDetails';
 
+import TestSuiteExecutionTabs from './TestSuiteExecutionTabs';
+
 const TestSuiteExecutionDrawer: FC = () => {
   const {close, id, data} = useExecutionDetailsPick('close', 'id', 'data');
   return (
     <ExecutionDrawer header={<ExecutionDrawerHeader />} loading={!data} open={Boolean(id)} onClose={close}>
-      <TestSuiteExecutionDetailsTabs />
+      <TestSuiteExecutionTabs />
     </ExecutionDrawer>
   );
 };
