+++ src/components/pages/TestSuites/TestSuiteDetails/TestSuiteExecution/TestSuiteExecutionTabs.tsx
@@ -10,7 +10,7 @@ import {useExecutionDetailsPick} from '@store/executionDetails';
 
 import {convertTestSuiteV2ExecutionToV3, isTestSuiteV2Execution} from '@utils/testSuites';
 
-const TestSuiteExecutionDetailsTabs: React.FC = () => {
+const TestSuiteExecutionTabs: React.FC = () => {
   const {data} = useExecutionDetailsPick('data') as {data: TestSuiteExecution};
 
   const testSuiteData = useMemo(
@@ -32,4 +32,4 @@ const TestSuiteExecutionDetailsTabs: React.FC = () => {
   );
 };
 
-export default TestSuiteExecutionDetailsTabs;
+export default TestSuiteExecutionTabs;
